Refactor main component into modules
